refactor: 셀을 이동하는 행위를 Coordinate 객체에 메세지를 보내도록 변경
